feat: Update to 26.1 (snapshot 1) (#320)
* chore: update used action versions and update java version

* chore: switch to lazy regex

* feat: switch to moddevgradle and update gradle to mc 26.1 and java 25

* fix: classtweaker path

* feat: update to 26.1 api

* chore: project-wide auto code cleanup

+++ common/src/main/java/com/klikli_dev/modonomicon/client/render/FluidBlockVertexConsumer.java
@@ -18,7 +18,7 @@ public record FluidBlockVertexConsumer(VertexConsumer prior, PoseStack pose, Blo
         final float dx = this.pos.getX() & 15;
         final float dy = this.pos.getY() & 15;
         final float dz = this.pos.getZ() & 15;
-        return this.prior.addVertex(this.pose.last().pose(), (float) x - dx, (float) y - dy, (float) z - dz);
+        return this.prior.addVertex(this.pose.last().pose(), x - dx, y - dy, z - dz);
     }
 
     @Override
